Updated Vagrantfile for trusty and other tweaks.
First and foremost this is an update to trusty rather than precise.  I switched
salt to its stable release since there did not seem to be daily builds for
trusty (at least not at that PPA).

I also changed the VM to use two CPUs and 1GB of memory, perhaps
unnecessariliy.

I added the ability to cache deb files.